decorators refactored cleanly
