+++ autoarray/structures/decorators/__init__.py
@@ -1,7 +1,7 @@
-from .to_array import grid_2d_to_array
-from .to_grid import grid_2d_to_grid
-from .to_vector_yx import grid_2d_to_vector_yx
-from .decorators import grid_1d_to_structure
-from .decorators import grid_1d_output_structure
-from .decorators import transform
-from .decorators import relocate_to_radial_minimum
+from .project_grid import project_grid
+from .to_projected import to_projected
+from .to_array import to_array
+from .to_grid import to_grid
+from .to_vector_yx import to_vector_yx
+from .transform import transform
+from .relocate_radial import relocate_to_radial_minimum
